Move helpers into utils folder
