+++ App/components/listItems/CredentialListItem.tsx
@@ -9,7 +9,7 @@ import { View, StyleSheet, TouchableOpacity } from 'react-native'
 import Icon from 'react-native-vector-icons/MaterialIcons'
 
 import { Colors } from '../../Theme'
-import { parseSchema } from '../../helpers'
+import { parseSchema } from '../../utils/helpers'
 import Text from '../texts/Text'
 import Title from '../texts/Title'
