+++ App/screens/ProofRequest.tsx
@@ -15,7 +15,7 @@ import { FlatList, Alert, View, StyleSheet } from 'react-native'
 import Toast from 'react-native-toast-message'
 
 import { Colors } from '../Theme'
-import { parseSchema } from '../helpers'
+import { parseSchema } from '../utils/helpers'
 
 import { Button, ModularView, Label } from 'components'
 import { HomeStackParams } from 'types/navigators'
